Move submitComment handler to seperate API file.
