+++ frontend/src/components/Comments/CommentsAPI.js
@@ -1,9 +1,20 @@
 export const CommentsAPI = {
-  getDivisions: () => {
-    return fetch('/divisions')
-      .then((res) => res.json())
-      .catch(function (error) {
-        console.log('Error fetching divisions', error);
-      });
+  submitComment: (formRef, articleID, commentID) => {
+    let commentToSubmit = {
+      author: 'ID? autor',
+      content: formRef.current.value,
+      articleID: articleID,
+      responseToCommentID: commentID,
+    };
+
+    const requestOptions = {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ commentToSubmit }),
+    };
+
+    fetch('/submit-comment', requestOptions).then((response) =>
+      response.json()
+    );
   },
 };
